Introduce Cells::Corner to reduce duplication;
This removes the duplicate #position method in some Cells classes.
