+++ lib/vedeu/cells/borders/corner.rb
@@ -1,0 +1,20 @@
+module Vedeu
+
+  module Cells
+
+    # Provides the position information to draw a border corner.
+    #
+    # @api private
+    #
+    class Corner < Vedeu::Cells::Border
+
+      # @return [Vedeu::Geometries::Position]
+      def position
+        @position = defaults[:position]
+      end
+
+    end # Corner
+
+  end # Cells
+
+end # Vedeu
